+++ NEWS.md
@@ -7,3 +7,7 @@
 - geom_arc / stat_arc for drawing circle segments
 - geom_edge_bundle / stat_edge_bundle for drawing edge bundles based on control
 points
+- power_trans for creating power transformations
+- radial_trans for creating transformation between radial and cartesian 
+coordinates
+-trans_reverser for reversing a trans object